+++ lib/scriptorium/index.ts
@@ -10,3 +10,6 @@
 
 export { renderConversationMarkdown } from './markdown-renderer'
 export { mergeAnnotations, stripAnnotations } from './annotation-merger'
+export type { ConversationMetadata } from '@/lib/schemas/scriptorium.types'
+export { searchConversationChunks } from './conversation-search'
+export type { ConversationSearchResult, ConversationSearchOptions } from './conversation-search'